+++ examples/table.rs
@@ -1,0 +1,47 @@
+use fltk::*;
+
+fn main() {
+    let app = app::App::default().set_scheme(app::AppScheme::Gtk);
+    let mut wind = window::Window::new(100, 100, 800, 600, "Table");
+    let mut table = table::Table::new(5, 5, 790, 590, "");
+
+    table.set_rows(30);
+    table.set_row_header(true);
+    table.set_row_resize(false);
+    table.set_cols(26);
+    table.set_col_header(true);
+    table.set_col_width_all(80);
+    table.set_col_resize(true);
+    table.end();
+
+    wind.end();
+    wind.show();
+
+    table.draw_cell(Box::new(move |ctx, row, col, x, y, w, h| match ctx {
+        table::TableContext::StartPage => draw::set_font(Font::Helvetica, 18),
+        table::TableContext::ColHeader => draw_header(&format!("{}", (col + 65) as u8 as char), x, y, w, h),
+        table::TableContext::RowHeader => draw_header(&format!("{}", row + 1), x, y, w, h),
+        table::TableContext::Cell => draw_data(&format!("{}", row + col), x, y, w, h),
+        _ => (),
+    }));
+
+    app.run().unwrap();
+}
+
+fn draw_header(s: &str, x: i32, y: i32, w: i32, h: i32) {
+    draw::push_clip(x, y, w, h);
+    draw::draw_box(FrameType::ThinUpBox, x, y, w, h, Color::FrameDefault);
+    draw::set_draw_color(Color::Black);
+    draw::draw_text2(s, x, y, w, h, Align::Center);
+    draw::pop_clip();
+}
+
+fn draw_data(s: &str, x: i32, y: i32, w: i32, h: i32) {
+    draw::push_clip(x, y, w, h);
+    draw::set_draw_color(Color::White);
+    draw::draw_rectf(x, y, w, h);
+    draw::set_draw_color(Color::Gray0);
+    draw::draw_text2(s, x, y, w, h, Align::Center);
+    draw::draw_rect(x, y, w, h);
+    draw::pop_clip();
+}