+++ fltk-sys/src/glu.rs
@@ -18675,9 +18675,3 @@ extern "C" {
         objW: *mut GLdouble,
     ) -> GLint;
 }
-extern "C" {
-    pub fn gl_start();
-}
-extern "C" {
-    pub fn gl_finish();
-}