+++ src/enums.rs
@@ -83,11 +83,11 @@ pub enum FrameType {
 #[repr(i32)]
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum Align {
-    AlignCenter = 0,
-    AlignTop = 1,
-    AlignBottom = 2,
-    AlignLeft = 4,
-    AlignRight = 8,
+    Center = 0,
+    Top = 1,
+    Bottom = 2,
+    Left = 4,
+    Right = 8,
 }
 
 /// Defines fonts used by FLTK
@@ -356,16 +356,43 @@ pub enum CallbackTrigger {
     EnterKeyChanged = 11,
 }
 
-/// Defines the Cursor styles supported by fltk
+/// Defines the text cursor styles supported by fltk
 #[repr(i32)]
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub enum CursorStyle {
+pub enum TextCursor {
     NormalCursor,
     CaretCursor,
     DimCursor,
     BlockCursor,
     HeavyCursor,
     SimpleCursor,
+}
+
+/// Defines the cursor styles supported by fltk
+#[repr(i32)]
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub enum Cursor {
+    Default = 0,
+    Arrow = 35,
+    Cross = 66,
+    Wait = 76,
+    Insert = 77,
+    Hand = 31,
+    Help = 47,
+    Move = 27,
+    NS = 78,
+    WE = 79,
+    NWSE = 80,
+    NESW = 81,
+    N = 70,
+    NE = 69,
+    E = 49,
+    SE = 8,
+    S = 9,
+    SW = 7,
+    W = 36,
+    NW = 68,
+    None = 255,
 }
 
 /// Defines the chart types supported by fltk