+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 ### Changes
 - (None)
 
+## [0.6.6] - 2020-06-17
+### Changes
+- RgbImage takes ownership of the underlying data.
+- Remove gl_start and gl_finish.
+- Add App::set_visual(Mode), app::own_colormap, app::set_focus.
+- Fix Cursor naming.
+- Update FLTK to commit 5005d04 for further SVG fixes.
+- Add TableExt::draw_cell and TableContext.
+- Rename Align enum members.
+
 ## [0.6.5] - 2020-06-16
 ### Changes
 - Fix cmake invocation on Windows to build for release.